refactor: 저장소 클래스 패키지 분리
- 별도의 저장소 패키지를 추가하고 이동합니다.
