+++ src/main/java/roomescape/application/reservation/ReservationService.java
@@ -11,13 +11,13 @@ import roomescape.application.reservation.dto.ReservationResult;
 import roomescape.application.reservation.dto.ReservationSearchParam;
 import roomescape.application.reservation.dto.ReservationWithStatusResult;
 import roomescape.domain.member.Member;
-import roomescape.domain.member.MemberRepository;
+import roomescape.domain.member.repository.MemberRepository;
 import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationRepository;
+import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.reservation.ReservationTime;
-import roomescape.domain.reservation.ReservationTimeRepository;
+import roomescape.domain.reservation.repository.ReservationTimeRepository;
 import roomescape.domain.reservation.Theme;
-import roomescape.domain.reservation.ThemeRepository;
+import roomescape.domain.reservation.repository.ThemeRepository;
 import roomescape.infrastructure.error.exception.MemberException;
 import roomescape.infrastructure.error.exception.ReservationException;
 import roomescape.infrastructure.error.exception.ReservationTimeException;
