+++ src/main/java/roomescape/application/reservation/WaitingService.java
@@ -6,14 +6,14 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.application.reservation.dto.CreateWaitingParam;
 import roomescape.application.reservation.dto.WaitingResult;
 import roomescape.domain.member.Member;
-import roomescape.domain.member.MemberRepository;
-import roomescape.domain.reservation.ReservationRepository;
+import roomescape.domain.member.repository.MemberRepository;
+import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.reservation.ReservationTime;
-import roomescape.domain.reservation.ReservationTimeRepository;
+import roomescape.domain.reservation.repository.ReservationTimeRepository;
 import roomescape.domain.reservation.Theme;
-import roomescape.domain.reservation.ThemeRepository;
+import roomescape.domain.reservation.repository.ThemeRepository;
 import roomescape.domain.reservation.Waiting;
-import roomescape.domain.reservation.WaitingRepository;
+import roomescape.domain.reservation.repository.WaitingRepository;
 import roomescape.domain.reservation.WaitingWithRank;
 import roomescape.infrastructure.error.exception.ReservationException;
