+++ src/main/java/roomescape/domain/member/repository/MemberRepository.java
@@ -1,7 +1,9 @@
-package roomescape.domain.member;
+package roomescape.domain.member.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import roomescape.domain.member.Email;
+import roomescape.domain.member.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
