+++ src/main/java/roomescape/domain/reservation/repository/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation;
+package roomescape.domain.reservation.repository;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import roomescape.domain.reservation.Reservation;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
