+++ src/main/java/roomescape/domain/reservation/repository/ThemeRepository.java
@@ -1,10 +1,11 @@
-package roomescape.domain.reservation;
+package roomescape.domain.reservation.repository;
 
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import roomescape.domain.reservation.Theme;
 
 public interface ThemeRepository extends JpaRepository<Theme, Long> {
