+++ src/main/java/roomescape/domain/reservation/repository/WaitingRepository.java
@@ -1,8 +1,10 @@
-package roomescape.domain.reservation;
+package roomescape.domain.reservation.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import roomescape.domain.reservation.Waiting;
+import roomescape.domain.reservation.WaitingWithRank;
 
 public interface WaitingRepository extends JpaRepository<Waiting, Long> {
