+++ src/test/java/roomescape/application/auth/AuthServiceTest.java
@@ -12,7 +12,7 @@ import roomescape.application.auth.dto.LoginParam;
 import roomescape.application.auth.dto.LoginResult;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
-import roomescape.domain.member.MemberRepository;
+import roomescape.domain.member.repository.MemberRepository;
 import roomescape.domain.member.Role;
 import roomescape.infrastructure.error.exception.LoginAuthException;
 import roomescape.infrastructure.security.JwtProperties;
