+++ src/test/java/roomescape/application/reservation/ReservationServiceTest.java
@@ -19,16 +19,15 @@ import roomescape.application.reservation.dto.ReservationWithStatusResult;
 import roomescape.application.reservation.dto.ThemeResult;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
-import roomescape.domain.member.MemberRepository;
+import roomescape.domain.member.repository.MemberRepository;
 import roomescape.domain.member.Role;
 import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationRepository;
+import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.reservation.ReservationStatus;
 import roomescape.domain.reservation.ReservationTime;
-import roomescape.domain.reservation.ReservationTimeRepository;
+import roomescape.domain.reservation.repository.ReservationTimeRepository;
 import roomescape.domain.reservation.Theme;
-import roomescape.domain.reservation.ThemeRepository;
-import roomescape.domain.reservation.WaitingRepository;
+import roomescape.domain.reservation.repository.ThemeRepository;
 import roomescape.infrastructure.error.exception.MemberException;
 import roomescape.infrastructure.error.exception.ReservationException;
 import roomescape.infrastructure.error.exception.ReservationTimeException;
